FOUND CAUSE AND FIXED
Using debug, I could see that rightSensor obtained a distance of -16 even when its yPosition is at 128 and it is seemingly on top of the fullTileBlock. Surely it should return 0?

What was happening at the corner to cause the jitter was the player briefly oscillated between the left foot touching the ground and the right foot touching the ground even if you moved a bit, as -16 was added and then a frame later this was corrected. This has been fixed by minusing 16-height instead of height