Issue: Added more checks, still get an ArrayIndexOutOfBoundsException.
Found out using a breakpoint that the check is still returning an error because it is checking all the conditions including those that involve e.g. map[chunkX] which causes an error when it is negative.

Also the player is now not attaching onto ledges correctly because of the change.